11205 - wip; ac1, moved ChartError.jsx, ChartLoadingMessage.jsx, ChartNoResults.jsx to sharedComponents, and changed file paths where they're used
